requirements.txt
pytube
Pillow
streamlit
langchain
replicate
pytesseract
transformers
python-dotenv
youtube_transcript_api
